+++ tSQLt_UnitTesting/testi21Database/Functions/test fnCalculateCostPerLot for the basics.sql
@@ -7,8 +7,8 @@ BEGIN
 			,@intLotUOMId AS INT
 			,@dblCostPerItemUOMId AS NUMERIC(18,6)
 
-	DECLARE @result AS NUMERIC(38,20)
-	DECLARE @expected AS NUMERIC(38,20)
+	DECLARE @result AS FLOAT
+	DECLARE @expected AS FLOAT
 
 	-- Act
 	SELECT @result = dbo.fnCalculateCostPerLot(
